u converter is now active

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jules <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/03 18:10:10 by jules             #+#    #+#             */
-/*   Updated: 2020/12/15 14:48:02 by jules            ###   ########.fr       */
+/*   Updated: 2020/12/15 15:16:12 by jules            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void		handle_flags(t_helper *helper, va_list *list, char c);
 void		pad_print(t_helper *helper);
 void		precision_print(t_helper *helper);
 
-int			count_digits(int val);
+int			count_digits(long val);
 
 void		handle_d(t_helper *helper, va_list *list);
